Bug fixed:
・Boolean型の変換に失敗する問題を解消。
・[RelatedUsers]のメールアドレス検索条件を精緻化。

+++ Implem.Pleasanter/Models/Issues/IssueModel.cs
@@ -2143,7 +2143,7 @@ namespace Implem.Pleasanter.Models
             {
                 if (notification.HasRelatedUsers())
                 {
-                    var users = new List<long>();
+                    var users = new List<int>();
                     Rds.ExecuteTable(
                         context: context,
                         statements: Rds.SelectIssues(
@@ -2158,10 +2158,10 @@ namespace Implem.Pleasanter.Models
                                 .AsEnumerable()
                                 .ForEach(dataRow =>
                                 {
-                                    users.Add(dataRow.Long("Manager"));
-                                    users.Add(dataRow.Long("Owner"));
-                                    users.Add(dataRow.Long("Creator"));
-                                    users.Add(dataRow.Long("Updator"));
+                                    users.Add(dataRow.Int("Manager"));
+                                    users.Add(dataRow.Int("Owner"));
+                                    users.Add(dataRow.Int("Creator"));
+                                    users.Add(dataRow.Int("Updator"));
                                 });
                     notification.ReplaceRelatedUsers(
                         context: context,
